PT-145/PT-148 Persist in pipeline, improve database connection

+++ db-layer/sql/generate.sql
@@ -40,7 +40,7 @@ INSERT INTO `event_type` VALUES (206, 'pr_remove_label');
 --
 DROP TABLE IF EXISTS `project`;
 CREATE TABLE `project` (
-  `_id`  INT(11)     NOT NULL AUTO_INCREMENT,
+  `_id`  BIGINT(11)     NOT NULL AUTO_INCREMENT,
   `name` VARCHAR(45) NOT NULL,
   PRIMARY KEY (`_id`),
   UNIQUE KEY `name_UNIQUE` (`name`)
@@ -51,7 +51,7 @@ CREATE TABLE `project` (
 --
 DROP TABLE IF EXISTS `repository`;
 CREATE TABLE `repository` (
-  `_id`     INT(11)      NOT NULL,
+  `_id`     BIGINT(11)      NOT NULL,
   `name`    VARCHAR(255) NOT NULL,
   `private` TINYINT(1)   NOT NULL,
   PRIMARY KEY (`_id`)
@@ -62,9 +62,9 @@ CREATE TABLE `repository` (
 --
 DROP TABLE IF EXISTS `project_repository`;
 CREATE TABLE `project_repository` (
-  `_id`           INT(11) NOT NULL AUTO_INCREMENT,
-  `project_id`    INT(11) NOT NULL,
-  `repository_id` INT(11) NOT NULL,
+  `_id`           BIGINT(11) NOT NULL AUTO_INCREMENT,
+  `project_id`    BIGINT(11) NOT NULL,
+  `repository_id` BIGINT(11) NOT NULL,
   PRIMARY KEY (`_id`),
   KEY `project_id_INDEX` (`project_id`),
   KEY `repository_id_INDEX` (`repository_id`),
@@ -81,7 +81,7 @@ CREATE TABLE `project_repository` (
 --
 DROP TABLE IF EXISTS `user`;
 CREATE TABLE `user` (
-  `_id`      INT(11)     NOT NULL,
+  `_id`      BIGINT(11)     NOT NULL,
   `username` VARCHAR(45) NOT NULL,
   PRIMARY KEY (`_id`),
   UNIQUE KEY `username_UNIQUE` (`username`)
@@ -92,10 +92,10 @@ CREATE TABLE `user` (
 --
 DROP TABLE IF EXISTS `event`;
 CREATE TABLE `event` (
-  `_id`            INT(11)   NOT NULL,
-  `repository_id`  INT(11)   NOT NULL,
-  `author_user_id` INT(11)   NOT NULL,
-  `owner_user_id`  INT(11)   NOT NULL,
+  `_id`            BIGINT(11)   NOT NULL,
+  `repository_id`  BIGINT(11)   NOT NULL,
+  `author_user_id` BIGINT(11)   NOT NULL,
+  `owner_user_id`  BIGINT(11)   NOT NULL,
   `event_type_id`  INT(11)   NOT NULL,
   `date`           TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP,
   PRIMARY KEY (`_id`),
